almost there

+++ lib/services/album_service.dart
@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:http/http.dart' as http;
 import 'dart:convert';
@@ -12,9 +13,9 @@ abstract class AlbumService {
   Stream<List<Album>> getCreatedAlbums();
   Stream<List<Album>> getJoinedAlbums();
   Stream<List<Photo>> getPhotos(String albumId);
+  Future<void> joinAlbum(String albumId);
   Future<void> createAlbum(String albumTitle);
-  Future<int> joinAlbum(String sharedString);
-  void deleteAlbum(String albumId, bool isOwner);
+  void deleteOrLeaveAlbum(String albumId, bool isOwner);
   Future<String> getUserProfileImg(String userId);
   Future<void> uploadImage(String imagePath, String albumId);
 }
@@ -98,33 +99,45 @@ class AlbumServiceImplementation implements AlbumService {
   }
 
   @override
-  void deleteAlbum(String albumId, bool isOwner) async {
-    FirebaseFirestore.instance.collection("albums").doc(albumId).delete();
+  void deleteOrLeaveAlbum(String albumId, bool isOwner) async {
+    if(isOwner) {
+      FirebaseFirestore.instance.collection("albums").doc(albumId).delete();
+    } else {
+      FirebaseFirestore.instance.collection("albums").doc(albumId).update({
+        'attendeeIds': FieldValue.arrayRemove([FirebaseAuth.instance.currentUser.uid])
+      });
+    }
+    
+  }
+
+  Future<bool> isUserInAlbum(String albumId) async {
+    bool isUserInAlbum = false;
+    String userId = FirebaseAuth.instance.currentUser.uid;
+    DocumentSnapshot data = await FirebaseFirestore.instance
+        .collection("albums")
+        .doc(albumId)
+        .get();
+    Album album = Album.fromJson(data.data());
+
+    if (album.attendeeIds.contains(userId) || album.ownerId == userId) {
+      isUserInAlbum = true;
+    }
+
+    return isUserInAlbum;
   }
 
   @override
-  Future<int> joinAlbum(String sharedString) async {
-    var client = http.Client();
-    try {
-      int userId = 42;
-      var url = host + "getAlbumIdFromShareString/$sharedString";
-      var response = await http.get(url);
-      var albumId = jsonDecode(response.body)[0]['album_id'];
-      Map body = {"albumId": albumId.toString(), "userId": userId.toString()};
-      var joinAlbumUrl = host + "joinAlbum";
-      var joinAlbumRes = await http.post(joinAlbumUrl, body: body);
-      if (joinAlbumRes.statusCode == 200) {
-        return 200;
-      } else if (joinAlbumRes.statusCode == 450) {
-        //User already in album
-        return 450;
-      } else {
-        return 400;
-      }
-    } catch (e) {
-      throw Exception("Error joing album");
-    } finally {
-      client.close();
+  Future<void> joinAlbum(String albumId) async {
+    String userId = FirebaseAuth.instance.currentUser.uid;
+    bool userInAlbum = await isUserInAlbum(albumId);
+
+    if (!userInAlbum) {
+      await FirebaseFirestore.instance
+          .collection("albums")
+          .doc(albumId)
+          .update({
+        "attendeeIds": [userId]
+      });
     }
   }
 }
